[FEAT] #16 더보기 페이지 위치 이동
